requirements.txt
torch>=2.0
torchvision>=0.15
numpy
pandas
opencv-python
Pillow
matplotlib
scikit-learn
scipy
scikit-image
tqdm
